requirements.txt
uv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-pinecone
langchain-groq
streamlit
python-dotenv
pypdf
torch==2.8.0 
torchvision==0.23.0